moved compiled C location from pulse/de to pulse/controllers
