+++ qiskit/providers/aer/pulse/controllers/digest_pulse_qobj.py
@@ -20,7 +20,7 @@ from collections import OrderedDict
 import numpy as np
 from qiskit.providers.aer.aererror import AerError
 # pylint: disable=no-name-in-module
-from ..de.pulse_utils import oplist_to_array
+from .pulse_utils import oplist_to_array
 
 
 class DigestedPulseQobj:
