+++ qiskit/providers/aer/pulse/controllers/unitary_controller.py
@@ -25,7 +25,7 @@ from qiskit.tools.parallel import parallel_map, CPU_COUNT
 from .pulse_de_solver import construct_pulse_zvode_solver
 
 # Imports from qutip_extra_lite
-from ..de.pulse_utils import occ_probabilities, write_shots_memory
+from .pulse_utils import occ_probabilities, write_shots_memory
 
 dznrm2 = get_blas_funcs("znrm2", dtype=np.float64)
